chore: do not track knowledge issues until we have elder_count
(If we're in prefix()... after that we track all knowledge issues).

We were seeing a few too many ProposeOffline calls during genesis ssection startup.
But it feels like we shouldn't need to track this in startup, where things are moving veryt fast
with join churn etc.
